requirements.txt
# Requirements for YourDaddy v3.1 - Advanced AI Assistant with Multilingual Support
# This file lists all Python packages required for the complete project

# Core AI/ML Agent Framework
autogen-agentchat==0.7.5
autogen-core==0.7.5
autogen-ext==0.7.5

# Google AI/Gemini Integration (with Vision Support)
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1

# Multilingual Support & Translation
# googletrans==4.0.0rc1      # Google Translate API - DISABLED: Conflicts with httpx
deep-translator==1.11.4    # Better translation library, supports multiple services
langdetect==1.0.9          # Language detection
# polyglot==16.7.4          # Multilingual NLP toolkit (optional) - DISABLED: Build issues
transliterate==1.10.2     # Text transliteration
indic-transliteration==2.3.37  # Indic scripts transliteration

# Voice Recognition & TTS (Multilingual)
SpeechRecognition==3.14.3  # Speech recognition library
pyttsx3==2.99              # Text-to-speech synthesis (fallback)
gTTS==2.5.3                # Google Text-to-Speech (online, good quality)
edge-tts==6.1.12           # Microsoft Edge TTS - Neural voices, best quality, no compilation
pygame>=2.5.2              # Audio playback for TTS
# azure-cognitiveservices-speech==1.38.0  # Azure Speech Services (optional) - DISABLED: Large download, optional

# Audio Processing
PyAudio==0.2.14            # Audio I/O for microphone access
pydub==0.25.1             # Audio manipulation
soundfile==0.12.1         # Audio file I/O
librosa==0.10.2           # Audio analysis

# Automation and Scheduling
schedule==1.2.2           # Task scheduling
APScheduler==3.11.0        # Advanced Python Scheduler

# Enhanced Learning & Machine Learning
scikit-learn==1.3.0       # Machine learning algorithms for pattern recognition
networkx==3.1             # Graph analysis for knowledge graphs
matplotlib==3.7.2         # Data visualization for insights

# Advanced System Integration
psutil==5.9.8             # System and process monitoring
watchdog==3.0.0           # File system monitoring
pywin32==311              # Windows API access (Windows only)
win32-setctime==1.1.0     # Windows file time manipulation

tensorflow==2.15.1  # For advanced AI models
torch==2.3.1  # PyTorch for AI models
torchvision==0.18.1  # Computer vision models
transformers==4.44.2  # Hugging Face transformers
sentence-transformers==3.0.1  # Semantic search and embeddings

# OpenAI Compatibility Layer
openai==2.8.0

# Windows Automation & GUI Control
pywinauto==0.6.9
pywin32-ctypes==0.2.3
pypiwin32==223
comtypes==1.4.13

# GUI Framework (Tkinter is built-in, but listing for completeness)
# tkinter - Built into Python standard library

# Offline Wake Word Detection
# pvporcupine==3.0.5  # DISABLED: Requires API key and additional setup

# Offline Speech-to-Text (STT) 
vosk==0.3.45

# Music Integration
spotipy==2.24.0  # Spotify Web API client library
ytmusicapi==1.8.1  # YouTube Music API integration

# Computer Vision & Image Processing
Pillow==11.3.0  # For screenshot capture and image manipulation
opencv-python==4.10.0.84  # Advanced computer vision
numpy==1.26.4  # Numerical computing for image processing
pyautogui==0.9.54  # Screenshot capture and GUI automation

# OCR and Document Processing
pytesseract==0.3.13  # Python wrapper for Tesseract OCR
PyPDF2==3.0.1  # PDF file manipulation
pdfplumber==0.11.4  # Advanced PDF text extraction

# Web Automation & Scraping
selenium==4.17.2  # Web browser automation
beautifulsoup4==4.12.3  # HTML/XML parsing for web scraping
lxml==4.9.4  # Fast XML and HTML parser
feedparser==6.0.11  # RSS/Atom feed parsing

# HTTP Client Libraries
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
urllib3==2.5.0

# Audio Processing (Text-to-Speech, Speech Recognition)
pyttsx3==2.99           # Text-to-speech engine
SpeechRecognition==3.14.3  # Online speech recognition (fallback)
PyAudio==0.2.14         # Audio I/O for microphone access

# Core Python packages
schedule==1.2.2
psutil==5.9.8

# Security & Authentication
Flask-JWT-Extended==4.6.0  # JWT authentication for Flask
Flask-Limiter==3.8.0       # Rate limiting for Flask
python-dotenv==1.0.1       # Environment variable management

# Notifications & Toast Messages
win10toast==0.9
plyer==2.1.0

# Web Framework for Modern UI
Flask==3.1.0
Flask-SocketIO==5.3.6
flask-cors==4.0.0
python-socketio==5.8.0
eventlet==0.33.3
Werkzeug==3.1.3

# Database Support (SQLite3 is built-in to Python)
# sqlite3 - Built into Python standard library

# Optional GUI Enhancements (for future features)
pygame==2.6.1           # Game development library (if needed)
tk==0.1.0               # Tkinter wrapper
tkcalendar==1.6.1       # Calendar widget
ttkbootstrap==1.16.0    # Modern themes for Tkinter
ttkthemes==3.2.2        # Additional Tkinter themes

# Data Processing & Validation
pydantic==2.12.4
pydantic_core==2.41.5

# Async/Concurrency Support
asyncio  # Built-in, no version needed
aiofiles==25.1.0
anyio==4.11.0

# Task Scheduling
APScheduler==3.11.0

# Utility Libraries
colorama==0.4.6
packaging==25.0
tqdm==4.67.1
regex==2025.11.3

# Protocol Buffers & gRPC
protobuf==5.29.5
grpcio==1.76.0
grpcio-status==1.71.2
googleapis-common-protos==1.72.0

# Deployment & Distribution (Optional)
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9

# Type Checking & Annotations
typing_extensions==4.15.0
annotated-types==0.7.0

# File and Data Processing
# json, base64, io, os, struct, threading, time - Built into Python standard library

# Core Python Libraries (Usually built-in but listed for completeness)
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
setuptools>=65.5.0

# Note: The following are built into Python and don't need separate installation:
# tkinter, sqlite3, json, base64, io, os, struct, threading, time, sys

# Additional Utilities
websockets==11.0.0
qrcode==7.4.0

# Performance Optimization
cachetools==5.3.0
aiofiles==25.1.0
memory_profiler==0.61.0
py-spy==0.3.0